+++ docs/CHANGE_LOG.md
@@ -1,7 +1,8 @@
 # Calendar.js - Change Log:
 
 ## Version 2.7.2:
-- 
+- The year jumping buttons available for the DatePicker mode can now be shown in the regular calendar mode.
+- Renamed the setting "addYearButtonsInDatePickerMode" to "addYearButtonsOnMainDisplay",
 
 <br>